Fikset problemer mellom GitHub og IntelliJ
>:(
