+++ src/main/java/graphql/dto/Place.java
@@ -1,6 +1,5 @@
-package graphql.model;
+package graphql.dto;
 
-import java.time.OffsetDateTime;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,11 +9,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-
-public class TripRequest {
-        Place from;
-        Place to;
-        OffsetDateTime dateTime;
-        Double walkSpeed;
-        Boolean arriveBy;
+public class Place {
+    String place;
 }
