+++ src/main/java/graphql/dto/TripPattern.java
@@ -1,6 +1,7 @@
-package graphql.model;
+package graphql.dto;
 
 import java.util.List;
+import java.time.OffsetDateTime;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,6 +11,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-public class TripResponse {
-        List<TripPattern> tripPatterns;
+public class TripPattern {
+        OffsetDateTime expectedRunTime;
+        OffsetDateTime expectedEndTime;
+        List<Leg> legs;
 }
