+++ src/main/java/graphql/dto/TripRequest.java
@@ -1,5 +1,6 @@
-package graphql.model;
+package graphql.dto;
 
+import java.time.OffsetDateTime;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,8 +10,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-public class Leg {
-        String mode;
-        Double distance;
-        Line line;
+
+public class TripRequest {
+    Place from;
+    Place to;
+    OffsetDateTime dateTime;
+    Double walkSpeed;
+    Boolean arriveBy;
 }
