+++ src/main/java/graphql/dto/TripResponse.java
@@ -1,7 +1,6 @@
-package graphql.model;
+package graphql.dto;
 
 import java.util.List;
-import java.time.OffsetDateTime;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -11,8 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-public class TripPattern {
-        OffsetDateTime expectedRunTime;
-        OffsetDateTime expectedEndTime;
-        List<Leg> legs;
+public class TripResponse {
+    List<TripPattern> tripPatterns;
 }
