+++ src/test/java/graphql/GraphqlApplicationTests.java
@@ -1,0 +1,13 @@
+package graphql;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GraphqlApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
